Test unmateched quotes
Note that all we will get in case of error is an error printout from
fish.  Since we cannot redirect this printout (fish does not allow it),
the test will not actually fail, as the return code will still be 0.

But getting an error printout is still better than nothing, in the hopes
of the problem being noticed.

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>