Avoid using newline as a completion separator
When calling the _completionTests_complete function in a subshell,
the trailing newlines are removed by the shell.  Therefore if the script
mistakenly includes an empty completion at the end, we would not detect
that bug.  See https://github.com/spf13/cobra/pull/1691

With this commit, we use a space to separate the completions, so an
empty one will be detected (extra space).

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>