fix(bash): Fix bash tests.
testprog was failing with "cannot execute: required file not found".
Some Google sleuthing found that it was probably because libc6 was
missing.

Redhat has a similar error where it was complaining that glibc v2.32 was
missing. ubi8 only seemed to be able to install v2.28, but ubi9 runs
testprog without any other tweaks.

Signed-off-by: Jeffrey Faer <jeffrey.faer@gmail.com>